[refactor] 상세 화면 패키지 구조 분리
- 병원, 메모 추가 관련 클래스를 각각의 패키지로 분리하고 이름을 간소화함.
