+++ mobile/app/src/main/java/com/lacuc/pets/MainModule.kt
@@ -2,9 +2,9 @@ package com.lacuc.pets
 
 import com.lacuc.pets.ui.animal.add.AddAnimalFragment
 import com.lacuc.pets.ui.animal.choose.ChooseAnimalFragment
-import com.lacuc.pets.ui.animal.detail.AnimalDetailAddMedicalFragment
-import com.lacuc.pets.ui.animal.detail.AnimalDetailAddMemoFragment
 import com.lacuc.pets.ui.animal.detail.AnimalDetailFragment
+import com.lacuc.pets.ui.animal.detail.medical.AddMedicalFragment
+import com.lacuc.pets.ui.animal.detail.memo.AddMemoFragment
 import com.lacuc.pets.ui.group.add.AddGroupFragment
 import com.lacuc.pets.ui.group.choose.ChooseGroupFragment
 import com.lacuc.pets.ui.login.signup.SignUpFragment
@@ -33,8 +33,8 @@ abstract class MainModule {
     abstract fun getAnimalDetailFragment(): AnimalDetailFragment
 
     @ContributesAndroidInjector
-    abstract fun getAnimalDetailAddMedicalFragment(): AnimalDetailAddMedicalFragment
+    abstract fun getAddMedicalFragment(): AddMedicalFragment
 
     @ContributesAndroidInjector
-    abstract fun getAnimalDetailAddMemoFragment(): AnimalDetailAddMemoFragment
+    abstract fun getAddMemoFragment(): AddMemoFragment
 }
