+++ mobile/app/src/main/java/com/lacuc/pets/ui/animal/detail/medical/AddMedicalFragment.kt
@@ -1,4 +1,4 @@
-package com.lacuc.pets.ui.animal.detail
+package com.lacuc.pets.ui.animal.detail.medical
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,21 +7,20 @@ import android.view.ViewGroup
 import androidx.fragment.app.viewModels
 import androidx.navigation.NavController
 import androidx.navigation.fragment.findNavController
+import androidx.navigation.ui.setupWithNavController
 import com.lacuc.pets.ViewModelFactory
-import com.lacuc.pets.databinding.FragmentAddMemoBinding
-import com.lacuc.pets.util.setupWithNavController
+import com.lacuc.pets.databinding.FragmentAddMedicalBinding
 import dagger.android.support.DaggerFragment
 import javax.inject.Inject
 
-class AnimalDetailAddMemoFragment : DaggerFragment() {
-
-    private var _binding: FragmentAddMemoBinding? = null
+class AddMedicalFragment : DaggerFragment() {
+    private var _binding: FragmentAddMedicalBinding? = null
     private val binding get() = _binding!!
 
     @Inject
     lateinit var viewModelFactory: ViewModelFactory
 
-    private val viewModel: AnimalDetailAddMemoViewModel by viewModels { viewModelFactory }
+    private val viewModel: AddMedicalViewModel by viewModels { viewModelFactory }
 
     private val navController: NavController by lazy { findNavController() }
 
@@ -30,7 +29,7 @@ class AnimalDetailAddMemoFragment : DaggerFragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentAddMemoBinding.inflate(inflater, container, false).apply {
+        _binding = FragmentAddMedicalBinding.inflate(inflater, container, false).apply {
             lifecycleOwner = viewLifecycleOwner
             vm = viewModel
         }
@@ -40,14 +39,14 @@ class AnimalDetailAddMemoFragment : DaggerFragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        binding.toolbarAddMemo.setupWithNavController(navController)
+        binding.toolbarAddMedical.setupWithNavController(navController)
 
         setOnCompleteEventObserver()
     }
 
     private fun setOnCompleteEventObserver() {
         viewModel.completeEvent.observe(viewLifecycleOwner) {
-            navController.previousBackStackEntry?.savedStateHandle?.set("onCompleteEvent", 2)
+            navController.previousBackStackEntry?.savedStateHandle?.set("onCompleteEvent", 1)
             navController.navigateUp()
         }
     }
